requirements.txt
PyYAML
-e git+git@git.corp.adobe.com:adorton/cce-umapi.git#egg=umapi
